Move particle mixins
